+++ packages/components/scripts/jobs/index.js
@@ -1,5 +1,7 @@
+export * from './angular.js';
 export * from './cem.js';
 export * from './createEventsFile.js';
+export * from './createExports.js';
 export * from './esbuildComponents.js';
 export * from './prepare.js';
 export * from './react.js';