+++ packages/components/scripts/vendorism.js
@@ -5,30 +5,10 @@ import commandLineArgs from 'command-line-args';
 import { eject, setSource, setTarget } from 'vendorism';
 import { optimizePathForWindows } from 'vendorism/src/scripts/helpers.js';
 import { generateStorybookFile, updateVsCodeReadOnlyFiles } from './vendorism/index.js';
-import {
-  vendorButton, vendorInput, vendorRadio, vendorTextarea,
-} from './vendorism/custom/index.js';
+import { vendorButton, vendorInput, vendorTextarea, vendorRadio } from './vendorism/custom/index.js';
+import { components, events } from './config.js';
 
-const components = [
-  'input',
-  'button',
-  'textarea',
-  'radio',
-];
-
-// List of events that we want to import.
-// @todo: Automate this depending on components!
-const events = [
-  'sl-blur',
-  'sl-focus',
-  'sl-invalid',
-  'sl-load',
-  'sl-error',
-  'sl-blur',
-  'sl-change',
-  'sl-clear',
-  'sl-input',
-].map(evt => `src/events/${evt}.ts`);
+const eventList = events.map(evt => `src/events/${evt}.ts`);
 
 const otherIncludes = [
   'custom-elements-manifest.config*',
@@ -40,7 +20,7 @@ const otherIncludes = [
   'src/shoelace-autoloader*',
   'src/translations/de.ts',
   'src/utilities/form*',
-  ...events,
+  ...eventList,
 ];
 
 const libraryPrefix = 'syn';
@@ -256,7 +236,7 @@ if (!options.getOnly) {
 
   const { removedFiles, newFiles } = await setTarget(config);
 
-  await updateVsCodeReadOnlyFiles(removedFiles, newFiles);
+  await updateVsCodeReadOnlyFiles(removedFiles, newFiles.filter(f => !f.includes('stories.ts')));
 
   // Move files back from './src/temp' to '../docs/src/components'
   await execSync('mv ./src/temp ../docs/stories/components');